fix: resolve exception filter error
