--- apps/server/src/common/filters/domain-exception.filter.ts
@@ -1,17 +1,0 @@
-import { ArgumentsHost, Catch, ExceptionFilter, HttpException } from '@nestjs/common';
-import { Response } from 'express';
-
-@Catch(HttpException)
-export class DomainExceptionFilter implements ExceptionFilter {
-  catch(exception: HttpException, host: ArgumentsHost) {
-    const ctx = host.switchToHttp();
-    const response = ctx.getResponse<Response>();
-
-    response.status(exception.getStatus()).json({
-      type: 'fail',
-      error: {
-        message: exception.message,
-      },
-    });
-  }
-}
